+++ Engine/src/function/impl/EnvironmentFunction.java
@@ -46,7 +46,7 @@ public class EnvironmentFunction extends AbstractFunction{
     }
 
     @Override
-    public Integer getTicksNotUpdated(Context context) {
+    public Float getTicksNotUpdated(Context context) {
         return null;
     }
