+++ Engine/src/function/impl/RandomFunction.java
@@ -45,7 +45,7 @@ public class RandomFunction extends AbstractFunction {
     }
 
     @Override
-    public Integer getTicksNotUpdated(Context context) {
+    public Float getTicksNotUpdated(Context context) {
         return null;
     }
 }
